+++ src/app/services/token/token.service.ts
@@ -14,12 +14,18 @@ export interface TokenResponseType {
   description: string;
   category: string;
   kind: string;
-  nonTransferable?: boolean;
-  booleanAmount?: boolean;
-  symbolPreference?: boolean;
+  isTransferable?: boolean;
+  isBooleanAmount?: boolean;
+  shouldPreferSymbol?: boolean;
+  series?: string;
 }
 export type ContractsType = Record<string, ContractType>;
 export type ContractType = FA12 | FA2;
+export type ContractsOverrideType = Record<string, ContractOverrideType>;
+export interface ContractOverrideType {
+  storageUsage?: number;
+  gasUsage?: number;
+}
 export interface TokensInterface {
   category: string;
 }
@@ -30,9 +36,10 @@ export interface TokenData {
   description: string;
   displayUrl: string;
   thumbnailUrl: string;
-  nonTransferable?: boolean;
-  booleanAmount?: boolean;
-  symbolPreference?: boolean;
+  isTransferable?: boolean;
+  isBooleanAmount?: boolean;
+  shouldPreferSymbol?: boolean;
+  series?: string;
 }
 export interface FA12 extends TokensInterface {
   kind: 'FA1.2';
@@ -50,7 +57,7 @@ export interface FA2 extends TokensInterface {
 
 export class TokenService {
   readonly AUTO_DISCOVER: boolean = true;
-  readonly version: string = '1.0.1';
+  readonly version: string = '1.0.7';
   private contracts: ContractsType = {};
   private exploredIds: Record<string, { firstCheck: number, lastCheck: number }> = {};
   readonly storeKey = 'tokenMetadata';
@@ -149,8 +156,9 @@ export class TokenService {
           description: metadata.description ? metadata.description : '',
           displayUrl,
           thumbnailUrl,
-          nonTransferable: metadata?.nonTransferable ? metadata.nonTransferable : false,
-          booleanAmount: metadata?.booleanAmount ? metadata.booleanAmount : false
+          isTransferable: metadata?.isTransferable ? metadata.isTransferable : true,
+          isBooleanAmount: metadata?.isBooleanAmount ? metadata.isBooleanAmount : false,
+          series: metadata.series ? metadata.series : undefined
         };
         contract.tokens[id] = token;
         this.addAsset(contractAddress, contract);
@@ -167,7 +175,7 @@ export class TokenService {
     } else {
       const token = this.exploredIds[tokenId];
       const timeout = (token.lastCheck - token.firstCheck) > 600000;
-      const reCheck = (now - token.lastCheck) > 2000;
+      const reCheck = (now - token.lastCheck) > 15000;
       if (timeout || !reCheck) {
         return false;
       }
@@ -229,8 +237,8 @@ export class TokenService {
     } else {
       const token = this.getAsset(tokenKey);
       if (token) {
-        if ((!token.symbolPreference && token.name) || !token.symbol) {
-          if (token.booleanAmount) {
+        if ((!token.shouldPreferSymbol && token.name) || !token.symbol) {
+          if (token.isBooleanAmount) {
             return `${token.name}`;
           } else {
             return `${Big(amount).div(10 ** (baseUnit ? token.decimals : 0)).toFixed()} ${token.name}`;