+++ src/environments/environment.prod.ts
@@ -7,13 +7,24 @@ export const CONSTANTS: Constants = {
   NAME: 'Mainnet',
   NETWORK: 'mainnet',
   MAINNET: true,
-  NODE_URL: 'https://mainnet-tezos.giganode.io',
+  NODE_URL: 'https://kukai-mainnet.tez.ie',
   BLOCK_EXPLORER_URL: 'https://tzkt.io',
-  CSI: {
-    url: 'https://conseil-prod.cryptonomic-infra.tech',
-    apiKey: 'klassare'
-  },
   ASSETS: {
+    'KT1K9gCRgaLRFKTErYt1wVxA3Frb9FjasjTV': { // kUSD
+      kind: 'FA1.2',
+      category: 'finance',
+      tokens: {
+        0: {
+          name: 'Kolibri USD',
+          symbol: 'kUSD',
+          decimals: 18,
+          description: 'Kolibri is a Tezos based stablecoin built on Collateralized Debt Positions (CDPs) known as Ovens.',
+          displayUrl: '../../../assets/img/tokens/kusd.png',
+          thumbnailUrl: '../../../assets/img/tokens/kusd.png',
+          shouldPreferSymbol: true
+        }
+      }
+    },
     'KT1LN4LPSqTMS7Sd2CJw4bbDGRkMv2t68Fy9': { // USDtz
       kind: 'FA1.2',
       category: 'finance',
@@ -25,7 +36,7 @@ export const CONSTANTS: Constants = {
           description: 'USDtz is a Tezos on-chain stablecoin pegged to the value of the United States Dollar.',
           displayUrl: '../../../assets/img/tokens/usdtz.png',
           thumbnailUrl: '../../../assets/img/tokens/usdtz.png',
-          symbolPreference: true
+          shouldPreferSymbol: true
         }
       }
     },
@@ -40,7 +51,7 @@ export const CONSTANTS: Constants = {
           description: 'tzBTC delivers the power of Bitcoin as a token on the Tezos blockchain.',
           displayUrl: '../../../assets/img/tokens/tzbtc.png',
           thumbnailUrl: '../../../assets/img/tokens/tzbtc.png',
-          symbolPreference: true
+          shouldPreferSymbol: true
         }
       }
     },
@@ -55,7 +66,7 @@ export const CONSTANTS: Constants = {
           description: 'ETHtz is Ethereum wrapped in the Tezos FA 2.0 token standard.',
           displayUrl: '../../../assets/img/tokens/ethtz.png',
           thumbnailUrl: '../../../assets/img/tokens/ethtz.png',
-          symbolPreference: true
+          shouldPreferSymbol: true
         }
       }
     },
@@ -70,7 +81,7 @@ export const CONSTANTS: Constants = {
           description: 'Wrapped Tezos by StakerDAO, a fully collateralized representation of XTZ.',
           displayUrl: '../../../assets/img/tokens/wxtz.png',
           thumbnailUrl: '../../../assets/img/tokens/wxtz.png',
-          symbolPreference: true
+          shouldPreferSymbol: true
         }
       }
     },
@@ -85,11 +96,42 @@ export const CONSTANTS: Constants = {
           description: 'This certificate verifies that the holder of its private key attended, contributed and completed the Tezos Israel and Madfish Solution Workshop on December 7th to the 9th, 2020. The certificate holder utilized skills in smart contract development and tokenization to build, test and deploy a token on the Tezos blockchain.',
           displayUrl: '../../../assets/img/tokens/mfil.jfif',
           thumbnailUrl: '../../../assets/img/tokens/mfil.jfif',
-          nonTransferable: true,
-          booleanAmount: true
+          isTransferable: false,
+          isBooleanAmount: true
+        }
+      }
+    },
+    'KT1AFA2mwNUMNd4SsujE1YYp29vd8BZejyKW': { // hDAO
+      kind: 'FA2',
+      category: 'finance',
+      tokens: {
+        0: {
+          name: 'hic et nunc DAO',
+          symbol: 'hDAO',
+          decimals: 6,
+          description: '',
+          displayUrl: '../../../assets/img/tokens/hdao.png',
+          thumbnailUrl: '../../../assets/img/tokens/hdao.png',
+          isTransferable: true,
+          isBooleanAmount: false,
+          shouldPreferSymbol: true
         }
       }
     }
+  },
+  CONTRACT_OVERRIDES: {
+    // hice et nunc
+    'KT1Hkg5qeNhfwpKW4fXvq7HGZB9z2EnmCCA9:mint_OBJKT': { storageUsage: 308 },
+    'KT1Hkg5qeNhfwpKW4fXvq7HGZB9z2EnmCCA9:swap': { storageUsage: 180 },
+    'KT1Hkg5qeNhfwpKW4fXvq7HGZB9z2EnmCCA9:curate': { storageUsage: 100 },
+    'KT1Hkg5qeNhfwpKW4fXvq7HGZB9z2EnmCCA9:collect': { storageUsage: 212 },
   }
 };
-export const TRUSTED_TOKEN_CONTRACTS = [];
+export const TRUSTED_TOKEN_CONTRACTS = [
+  'KT1Em3sjKdHo3Fo9Az4EusZXQbkgsdZHkQkF',
+  'KT1RJ6PbjHpwc3M5rw5s2Nbmefwbuwbdxton',//hicetnunc
+  'KT1M2JnD1wsg7w2B4UXJXtKQPuDUpU2L7cJH',//hicetnunc-legacy
+  'KT1W4wh1qDc2g22DToaTfnCtALLJ7jHn38Xc',//alchememist
+  'KT1EH8yKXkRoxNkULRB1dSuwhkKyi5LJH82o',//mandala
+  'KT1DKBvxiwDR7qazNuuxCxY2AaXnoytmDE7H',//Mandala v2
+];